requirements.txt
asgiref==3.11.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security==5.7.1
Flask-SQLAlchemy==3.1.1
Flask-SQLAlchemy-Lite==0.2.1
Flask-WTF==1.2.2
greenlet==3.3.2
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.3
MarkupSafe==3.0.3
packaging==26.0
python-dotenv==1.2.2
requests==2.32.5
setuptools==82.0.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
WTForms==3.2.1
